Game: Jumping logic redesigned (and now on SPACE key), simple Break (DOWN), simple SlowMotion ('S'), little fixes and optimizations; Editor: You can now move polygons using LMB and add polygon points using RMB

+++ trunk/Bounce/GameState.h
@@ -27,7 +27,7 @@ public:
 	virtual int init(Application* application) = 0;
 	virtual void destroy() = 0;
 	virtual void processEvents() = 0;
-	virtual void processLogic(double time) = 0;
+	virtual void processLogic(float time) = 0;
 	virtual void processGraphics() = 0;
 	virtual ~GameState() = 0 {}
 };
